Update ng_dwi_recon_fncs.py
Fixed a bug in the WLSQ implementation. For WLSQ, want to solve

$\min_X ||W^{1/2} (Y - EX)||_2$

Instead of solving the normal Equation 

$E^HWY = E^HWEX$,

the code was incorrectly solving

$\min_X || W^{1/2} (W{-1/2}Y - EX))||_2$ by solving the normal equation

$E^HW^{1/2}Y = E^HWEX$

However, this only changes the signal levels by a few percent likely because the weights lose their spatially varying nature when applied to multi-shot data where each shot is not Nyquist sampled.  

For example if you choose a 1D line, most weights may be close to 1. If there is a region that is trying to be rejected, then this low weight region is averaged with the regular data. You can also try to play around with different methods of calculating weights (for example squaring all weights) and solving the WLSQ problem, and find that it only makes a few percent signal difference.